Refactor getReplacementPlan
