+++ src/main/scala/com/microsoft/hyperspace/index/rules/RuleUtils.scala
@@ -18,10 +18,14 @@ package com.microsoft.hyperspace.index.rules
 
 import scala.collection.mutable
 
+import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
-import org.apache.spark.sql.execution.datasources.LogicalRelation
+import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, RepartitionByExpression}
+import org.apache.spark.sql.execution.datasources.{HadoopFsRelation, InMemoryFileIndex, LogicalRelation}
+import org.apache.spark.sql.execution.datasources.parquet.ParquetFileFormat
+import org.apache.spark.sql.types.StructType
 
+import com.microsoft.hyperspace.BucketUnionLogicalPlan
 import com.microsoft.hyperspace.actions.Constants
 import com.microsoft.hyperspace.index.{IndexConstants, IndexLogEntry, IndexManager, LogicalPlanSignatureProvider}
 
@@ -80,6 +84,95 @@ object RuleUtils {
       Some(lrs.head)
     } else {
       None // logicalPlan is non-linear or it has no LogicalRelation.
+    }
+  }
+
+  /**
+   * Get replacement plan for current plan. The replacement plan reads data from indexes
+   *
+   * Pre-requisites
+   * - We know for sure the index which can be used to replace the plan.
+   *
+   * NOTE: This method currently only supports replacement of Scan Nodes i.e. Logical relations
+   *
+   * @param index index used in replacement plan
+   * @param plan current plan
+   * @return replacement plan
+   */
+  def getReplacementPlan(
+      spark: SparkSession,
+      index: IndexLogEntry,
+      plan: LogicalPlan): LogicalPlan = {
+    val location = new InMemoryFileIndex(spark, Seq(new Path(index.content.root)), Map(), None)
+    val relation = HadoopFsRelation(
+      location,
+      new StructType(),
+      index.schema,
+      Some(index.bucketSpec),
+      new ParquetFileFormat,
+      Map())(spark)
+
+    // Here we can't replace the plan completely with the index. This will create problems.
+    // For e.g. if Project(A,B) -> Filter(C = 10) -> Scan (A,B,C,D,E)
+    // if we replace this plan with index Scan (A,B,C), we lose the Filter(C=10) and it will
+    // lead to wrong results. So we only replace the base relation.
+
+    val replacedPlan = plan transformUp {
+      case baseRelation @ LogicalRelation(_: HadoopFsRelation, baseOutput, _, _) =>
+        val updatedOutput =
+          baseOutput.filter(attr => relation.schema.fieldNames.contains(attr.name))
+        baseRelation.copy(relation = relation, output = updatedOutput)
+    }
+
+    val hybridScanEnabled = spark.sessionState.conf
+      .getConfString(
+        IndexConstants.INDEX_HYBRID_SCAN_ENABLED,
+        IndexConstants.INDEX_HYBRID_SCAN_ENABLED_DEFAULT.toString)
+      .toBoolean
+
+    if (hybridScanEnabled) {
+      getComplementIndexPlan(spark, index, plan, replacedPlan)
+    } else {
+      replacedPlan
+    }
+  }
+
+  private def getComplementIndexPlan(
+      spark: SparkSession,
+      index: IndexLogEntry,
+      originalPlan: LogicalPlan,
+      indexPlan: LogicalPlan): LogicalPlan = {
+    val complementIndexPlan = originalPlan transformUp {
+      case baseRelation @ LogicalRelation(fsRelation: HadoopFsRelation, baseOutput, _, _) =>
+        val updatedOutput =
+          baseOutput.filter(attr => index.schema.fieldNames.contains(attr.name))
+
+        val files =
+          fsRelation.location.inputFiles
+            .map(new Path(_))
+            .filter(!index.allSourceFileSet.contains(_))
+
+        if (files.nonEmpty) {
+          val newLocation =
+            new InMemoryFileIndex(spark, files, Map(), None)
+          val newRelation =
+            fsRelation.copy(location = newLocation, dataSchema = index.schema)(spark)
+          baseRelation.copy(relation = newRelation, output = updatedOutput)
+        } else {
+          baseRelation
+        }
+    }
+
+    if (!originalPlan.equals(complementIndexPlan)) {
+      val attrs = complementIndexPlan.output.attrs.filter { attr =>
+        index.indexedColumns contains attr.name
+      }
+      val shuffled = RepartitionByExpression(attrs, complementIndexPlan, index.numBuckets)
+      BucketUnionLogicalPlan(
+        Seq(indexPlan, shuffled),
+        index.bucketSpec.copy(sortColumnNames = Seq()))
+    } else {
+      indexPlan
     }
   }
 }
